Updated Predict and sorted out functionality elsewhere

+++ twitter_classifier/scala/src/main/scala/com/databricks/apps/twitterClassifier/ExamineAndTrainOptions.scala
@@ -4,25 +4,24 @@ import java.io.File
 import com.github.acrisci.commander.Program
 
 abstract sealed case class ExamineAndTrainOptions(
-  twitterOptions: TwitterOptions,
   overWrite: Boolean = false,
-  tweetDirectory: File = new File("~/twitterClassifier/tweets/"),
-  modelDirectory: File = new File("~/twitterClassifier/modelDirectory/"),
+  tweetDirectory: File = new File("~/sparkTwitter/tweets/"),
+  modelDirectory: File = new File("~/sparkTwitter/modelDirectory/"),
   numClusters: Int = 10,
   numIterations: Int = 100
 )
 
-object ExamineAndTrainOptions extends CommonOptions {
-  override val _program = super._program
-    .option(flags="-x, --overWrite", description="Overwrite data files on each run [false]", default=false)
-    .epilogue("<tweetDirectory> <modelDirectory> <numClusters> <numIterations>")
+object ExamineAndTrainOptions {
+  val _program = new Program()
+    .version("2.0.0")
+    .option(flags="-w, --overWrite", description="Overwrite model from a previous run [false]", default=false)
+    .usage("ExamineAndTrain [Options] <tweetDirectory> <modelDirectory> <numClusters> <numIterations>")
 
   def parse(args: Array[String]): ExamineAndTrainOptions = {
     val program: Program = _program.parse(args)
-    if (program.args.length!=program.epilogue.split(" ").length) program.help
+    if (program.args.length!=program.usage.split(" ").length-2) program.help
 
     val options = new ExamineAndTrainOptions(
-      twitterOptions = super.apply(args),
       overWrite = program.overWrite,
       tweetDirectory = new File(program.args.head),
       modelDirectory = new File(program.args(1)),
@@ -36,9 +35,14 @@ object ExamineAndTrainOptions extends CommonOptions {
       System.exit(-1)
     }
     if (modelDirectory.exists) {
-      import org.apache.commons.io.FileUtils
-      println("Replacing pre-existing model")
-      FileUtils.deleteDirectory(modelDirectory)
+      if (options.overWrite) {
+        import org.apache.commons.io.FileUtils
+        println("Replacing pre-existing model")
+        FileUtils.deleteDirectory(modelDirectory)
+      } else {
+        System.err.println("Model already exists and --overWrite option was not specified")
+        System.exit(-3)
+      }
     }
     if (numClusters<1) {
       System.err.println(s"At least 1 clusters must be specified")
